backend: reorganize containerEngineClient files
