+++ pkg/rancher-desktop/backend/containerClient/index.ts
@@ -1,0 +1,3 @@
+export * from './types';
+export { MobyClient } from './mobyClient';
+export { NerdctlClient } from './nerdctlClient';
