+++ pkg/rancher-desktop/backend/wsl.ts
@@ -15,10 +15,8 @@ import {
   BackendError, BackendEvents, BackendProgress, BackendSettings, execOptions, FailureDetails, RestartReasons, State, VMBackend, VMExecutor,
 } from './backend';
 import BackendHelper from './backendHelper';
-import { ContainerEngineClient } from './containerEngine';
+import { ContainerEngineClient, MobyClient, NerdctlClient } from './containerClient';
 import K3sHelper from './k3sHelper';
-import MobyClient from './mobyClient';
-import NerdctlClient from './nerdctlClient';
 import ProgressTracker, { getProgressErrorDescription } from './progressTracker';
 
 import DEPENDENCY_VERSIONS from '@pkg/assets/dependencies.yaml';
